requirements.txt
asgiref==3.7.2
charset-normalizer==3.2.0
Django==4.2.5
django-phonenumber-field==7.1.0
phonenumberslite==8.13.21
Pillow==10.0.1
pypugjs==5.9.12
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
